requirements.txt
requests
pandas
gspread
gspread-dataframe
websocket-client
google-auth 
python-dotenv
